requirements.txt
fastmcp>=2.14.1
aiohttp>=3.13.0
spacy>=3.8.0
wikipedia>=1.4.0
matplotlib>=3.5.0
pandas>=1.3.0
pytest>=7.0.0
seaborn>=0.11.0
numpy>=1.21.0
weasyprint>=62.0
pdfkit>=1.0.0